+++ src/shape.h
@@ -94,7 +94,7 @@ b3AABB b3ComputeSweptShapeAABB( const b3Shape* shape, const b3Sweep* sweep, floa
 b3Vec3 b3GetShapeCentroid( const b3Shape* shape );
 float b3GetShapeArea( const b3Shape* shape );
 float b3GetShapeProjectedArea( const b3Shape* shape, b3Vec3 planeNormal );
-uint64_t b3GetShapeUserMaterialId( const b3Shape* shape, int childIndex, int triangleIndex );
+uint32_t b3GetShapeUserMaterialId( const b3Shape* shape, int childIndex, int triangleIndex );
 
 b3ShapeProxy b3MakeShapeProxy( const b3Shape* shape );
 b3ShapeProxy b3MakeLocalProxy( const b3ShapeProxy* proxy, b3Transform transform, b3Vec3* buffer );
